WebIDL/tests/submissions/heycam/double-003.html

<!DOCTYPE html>
<title>double</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.14 double
//
// IDL double values are represented with finite JS Number values.

// feature: HTMLProgressElement.value (attribute) (of type double)

test(function() {
  var e = document.createElement("progress");
  assert_equals(typeof e.value, "number");
  assert_true(e.value * 0 == 0);
}, "HTMLProgressElement.value returns a value of the correct type");

test(function() {
  var e = document.createElement("meter");
  assert_equals(typeof e.min, "number");
  assert_true(e.min * 0 == 0);
}, "HTMLMeterElement.min returns a value of the correct type");
</script>
